(release_30)bugfix: wrong ID number Data-type

+++ src/AliasUnit.cpp
@@ -258,7 +258,7 @@ void AliasUnit::removeAlias( TAlias * pT )
 }
 
 
-qint64 AliasUnit::getNewID()
+int AliasUnit::getNewID()
 {
     return ++mMaxID;
 }
