+++ src/Tree.h
@@ -45,8 +45,8 @@ public:
     void               Dump();
     void               setFullyExpanded()          { FullyExpanded = true; }
     bool               isFullyExpanded()           { return FullyExpanded; }
-    qint64             getID()                     { return mID; }
-    void               setID( qint64 id )          { mID=id; }
+    int                getID()                     { return mID; }
+    void               setID( int id )             { mID=id; }
     void               addChild( T * newChild, int parentPostion = -1, int parentPosition = -1 );
     bool               popChild( T * removeChild );
     void               setParent( T * parent );
@@ -62,7 +62,7 @@ public:
 
     T *                mpParent;
     std::list<T *> *   mpMyChildrenList;
-    qint64             mID;
+    int                mID;
     QString &          getError();
     void               setError( QString );
     bool               state();
